requirements.txt
aiogram>=3.0
openai
python-dotenv
mysql-connector-python